requirements.txt
Flask
google-generativeai
PyPDF2
docx2txt
werkzeug
groq
tiktoken
pdfplumber